Added Steffi's fix for revassign test 003 to prevent nil pointer deref

+++ testdata/reverseassign/003-binary_expression_spacing/main.go
@@ -1,11 +1,11 @@
-// <<<<< reverseassign,8,3,8,10,fail
+// <<<<< reverseassign,8,3,8,21,fail
 package main
 
 import "fmt"
 import "reflect"
 
 func main() {
-  var i int=3+4.5
+  var i float64=3+4.5
   fmt.Println("type of i is", reflect.TypeOf(i))
   fmt.Println("value of i is", i)
 }
